fix clipping issue with hug widgets in an alignment
